### 2018-08-27 #908 @Kevin-Mattheus-Moerman
@danielhatton thanks for your submission.
One of the JOSS [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) is:

> The software should have an obvious research application

It is clear to me that this work offers a useful means to automate the creation of 2D drawings for FreeCAD 3D models (which are BS 8888:2011 compliant). However can you elaborate on the use of the software in research, and how it may influence research outcomes.
