### 2018-09-11 3b4b695
Update paper.bib
