+++ paper/paper.bib
@@ -80,3 +80,45 @@ pages="151--159",
 doi="10.1007/s12008-008-0045-5",
 location="URL: <https://doi.org/10.1007/s12008-008-0045-5>"
 }
+
+@ARTICLE{Magee:2012:CIS,
+author="Justin Magee and Brian McClelland and John Winder",
+title="Current issues with standards in the measurement and documentation of human skeletal anatomy",
+journal="J. Anat.",
+year=2012,
+month=sep,
+volume=221,
+number=3,
+pages="240--251",
+doi="10.1111/j.1469-7580.2012.01535.x",
+location="URL: <https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1469-7580.2012.01535.x>"
+}
+
+@INPROCEEDINGS{Moitinho_de_Almeida:2014:MDT,
+author="Moitinho de Almeida, Vera and Juan Antonio Barcel{\'o}",
+title="Measuring and Describing {3D} Texture",
+pages="519--528",
+address="Paris",
+year=2014,
+organization="CAA International",
+month=apr # "\ 23",
+editor="F. Giligny and F. Djindjian and L. Costa and P. Moscati and S. Robert",
+booktitle="Proceedings of the 42nd Annual Conference on Computer Applications and Quantitative Methods in Archaeology {CAA} 2014 --- 21st Century Archaeology",
+publisher="Archaeopress",
+isbn="9781784911003",
+isbn="9781784911010"
+}
+
+@BOOK{Shapin:2011:LAP,
+author="Steven Shapin and Simon Schaffer",
+isbn="9780691150208",
+publisher="Princeton University Press",
+title="Leviathan and the Air-Pump: {H}obbes, {B}oyle, and the Experimental Life.},
+location= "URL: <http://search.ebscohost.com.plymouth.idm.oclc.org/login.aspx?direct=true&AuthType=ip,url,shib&db=nlebk&AN=390526&site=ehost-live>",
+year=2011,
+address="Princeton",
+series="Princeton Classics",
+edition="Paperback reissue, with a new introduction",
+note="First edition published 1985"
+}
+
