+++ paper/paper.bib
@@ -113,7 +113,7 @@ isbn="9781784911010"
 author="Steven Shapin and Simon Schaffer",
 isbn="9780691150208",
 publisher="Princeton University Press",
-title="Leviathan and the Air-Pump: {H}obbes, {B}oyle, and the Experimental Life.},
+title="Leviathan and the Air-Pump: {H}obbes, {B}oyle, and the Experimental Life.",
 location= "URL: <http://search.ebscohost.com.plymouth.idm.oclc.org/login.aspx?direct=true&AuthType=ip,url,shib&db=nlebk&AN=390526&site=ehost-live>",
 year=2011,
 address="Princeton",
