### 2018-09-11 #908 @danielhatton
Dear Dr. Moerman,

I've added a paragraph to the statement of need in this paper, suggesting the applicability of the tool in research contexts.  Does it help?
