### 2018-09-23 #974 @whedon
[REVIEW]: eights: BS 8888:2011 first angle projection drawings from FreeCAD 3D model
**Submitting author:** @danielhatton (<a href="http://orcid.org/0000-0002-4585-3336">Daniel Hatton</a>)
**Repository:** <a href="https://github.com/danielhatton/eights/" target ="_blank">https://github.com/danielhatton/eights/</a>
**Version:** 0.0
