### 2018-09-24 danielhatton/eights#1 @devinberg
Missing LICENSE file
*Part of the peer-review for JOSS: https://github.com/openjournals/joss-reviews/issues/974*

The [JOSS criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license) specify that the software license should be located in a file called `LICENSE`. Currently this information is contained within `COPYING.txt` and `COPYING_ANNEX.txt`.
