### 2018-09-24 danielhatton/eights#2 @devinberg
Documentation missing Statement of Need
*Part of the peer-review for JOSS: https://github.com/openjournals/joss-reviews/issues/974*

According to the [JOSS criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need), the software's documentation, found in `Documentation.txt`, should include a Statement of Need indicating "what problems the software is designed to solve and who the target audience is." This is partially addressed in the DESCRIPTION paragraph of the documentation file. However, that section is missing information on target audience.
