### 2018-09-24 danielhatton/eights#3 @devinberg
Missing Community Guidelines
*Part of the peer-review for JOSS: https://github.com/openjournals/joss-reviews/issues/974*

The [JOSS criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) specify that the repo should contain "Contributing guidelines" which provide direction for anyone wishing to contribute, report issues, or seek support. This information might normally be found in a file called `CONTRIBUTING`.
