### 2018-09-24 danielhatton/eights#4 @devinberg
Installation and testing
*Part of the peer-review for JOSS: https://github.com/openjournals/joss-reviews/issues/974*

I've been working through the installation and documentation. I was able to use `eights.py` by placing it in the macros directory. However, I was not able to get it to work by saving the file in the working directory.

Could you help me better understand what I should be seeing as output?

I was able to get a drawing to generate (see [here](https://github.com/danielhatton/eights/files/2412779/Two_parts.pdf)) using `Two_parts.py` but not with `Single_part.py` or `Assembly.py`.
