### 2018-10-16 3f9651a
Update eights.py
