### 2018-10-16 danielhatton/eights#1 @danielhatton
Thanks for the heads-up, @devinberg.  I believe this issue to be resolved by the latest set of commits on the master branch.
