+++ Documentation.txt
@@ -50,6 +50,70 @@ automates the construction of a page of 2D axonometric drawings in
 first angle projection, in a style consistent (to the best of the
 module author's ability) with the BS 8888:2011 standard.
 
+STATEMENT OF NEED
+
+The creative process behind detailed engineering design now typically takes place in 3D
+CAD packages (Quintana, Rivest, Pellerin, Venne, & Kheddouci, 2010). However, for
+this creative process actually to lead to an embodiment of the design being manufac-
+tured, this creative process must be followed by communication of the content of the
+design to a manufacturing facility. There are two necessary requirements for success:
+firstly, the communication must include information on all those aspects of the design
+that are necessary to complete manufacture, and to verify by inspection that the com-
+pleted, manufactured artefact matches the design (Quintana et al., 2010); secondly, the
+communication must be in a well-defined (usually graphical) language, so that it can be
+understood at the manufacturing facility in a way that permits successful manufacturing
+and inspection (Dobelis et al., 2018; Quintana et al., 2010). The drive to meet both of
+these requirements has led to the development of published standards for design commu-
+nication (Dobelis et al., 2018), of which BS 8888:2011 (Technical product documentation
+and specification BS 8888, 2011) is one. These standards are in a very advanced state of
+maturity for 2D engineering drawings, but despite efforts in recent years to develop sim-
+ilar standards for 3D models, the standards for 3D models remain somewhat less mature
+(Quintana et al., 2010). As a result, 2D drawings remain a crucial medium for commu-
+nication of design information to manufacturing facilities, and possessing the capability
+for automated generation of standards-compliant 2D drawings from a 3D model is an im-
+portant criterion by which the technical quality of 3D CAD software packages is assessed
+(Hughes, 2013; Kannan & Vinay, 2008). Hence, by automating the process of producing
+certain standards-compliant 2D drawings, the eights module offers the opportunity for the
+FreeCAD CAD package (FreeCAD version 0.15 user manual, n.d.) to be more favourably
+assessed against this criterion in future.
+
+One automated tool, for generating 2D drawings in first angle projection from a 3D model
+in FreeCAD, already exists: the FreeCAD Automatic Drawing Macro (“Macro automatic
+drawing,” 2016). A full analysis of the relative advantages and disadvantages of the
+Automatic Drawing Macro, and the eights module announced here, can be found in the
+"COMPARISON AND CONTRAST WITH OTHER SOFTWARE OF SIMILAR PURPOSE" section of this
+documentation file.  For the purposes of this statement of need, it is
+sufficient to mention two of the relative disadvantages of the Automatic Drawing Macro:
+-  the Automatic Drawing Macro does not attempt to comply with BS 8888 as regards
+   the format of the drawing sheet and its title block;
+-  the Automatic Drawing Macro does not offer the ability to include first angle pro-
+   jection sets for multiple 3D objects on the same drawing sheet.
+
+It is not only in the manufacturing sector that standardized axonometric drawings have
+proved important to clear communication: scholarly research has also benefited from this
+language. For example, in the study of human anatomy, standardization of axonometric
+drawing sets to the BS 8888 standard has been used to reduce both cumulative uncertainty
+in position co-ordinates and difficulty of interpretation associated with the presentation,
+in published papers, of geometric data on the human skeleton (Magee, McClelland, &
+Winder, 2012); and in archaeology, the standard first-angle projection set has been used
[…]
