### 2018-10-16 danielhatton/eights#2 @danielhatton
A "statement of need" section, largely copied from the JOSS manuscript, has now been added to the "Documentation.txt" file.  I will, however, leave this issue open, pending @devinberg's confirmation that the added section adequately resolves the issue.
