### 2018-10-16 danielhatton/eights#3 @danielhatton
Thanks, @devinberg.  I'll familiarize myself with what forms of collaborative editing are possible (or can be made possible by adjusting permissions settings) on GitHub, and compose something appropriate.  It may take a while.....
