### 2018-10-16 danielhatton/eights#4 @danielhatton
Can I have some further information to help debug, please?  In particular:

- On what operating system and FreeCAD version were you testing (I'm currently on FreeCAD 0.16 for Scientific Linux 7.5, and it works for me)?

- How did you launch FreeCAD in such a way as to give it a particular working directory (I'm currently launching FreeCAD from a bash command line with the directory in question as working directory, and it works for me)?

- Did you run Single_part.py and/or Assembly.py from a freshly-launched FreeCAD instance, or from a FreeCAD instance that already had some documents open from a previous run of Two_parts.py (it works for me in the first case, but not the second)?

I've made copies (via "print to file" in the FreeCAD GUI) of the outputs I get from Single_part.py and Assembly.py, but every time I try to attach them to this comment, I get a message from GitHub saying 'Something went really wrong, and we can't process that file'.  I'll try it again tomorrow, from a different network environment.
