### 2018-10-17 danielhatton/eights#4 @devinberg
I've tried this on Windows 10 with both FreeCAD 0.17 and 0.15. Version 0.17 doesn't appear to work completely as intended, but 0.15 I think does. However, I'd like to see your output so I know if I'm getting the correct thing.

In general, I think your documentation would be improved by providing this sort of benchmark to allow users to make sure the software is working as intended on their system. The JOSS criteria describes automated tests which might not make sense in this case. However, I think you could take your example documents and provide those with the intended output as a way of helping users test their system.

This might be easier to do if you migrated your `documentation.txt` file to a `README.md` file. Then you could include some screenshots showing the intended output for a test input, providing users a comparison point.
