### 2018-10-17 danielhatton/eights#3 @devinberg
An example of a relatively minimal contributing document can be found here: https://gitlab.com/moorepants/skijumpdesign/blob/master/CONTRIBUTING
