### 2018-10-18 danielhatton/eights#4 @danielhatton
Here are the intended outputs from Single_part.py and Assembly.py:
[Single_part.pdf](https://github.com/danielhatton/eights/files/2492355/Single_part.pdf)

[Assembly.pdf](https://github.com/danielhatton/eights/files/2492353/Assembly.pdf)

Yes, that's a good idea about including them in the project tree for benchmarking.
