### 2018-10-18 danielhatton/eights#4 @devinberg
@danielhatton Thanks for providing those. Only difference I'm seeing when I run the examples is the positioning of the text in the title block. See below where the text is double spaced and the info on sheet size, part number and such has moved to different boxes. Any ideas what would be causing that?

I'm running version 0.15 and using the example code you provide.

![assembly_title_block](https://user-images.githubusercontent.com/1399027/47175462-c71aee80-d2d8-11e8-8054-368261fbda37.PNG)
