### 2018-10-24 #974 @devinberg
@Kevin-Mattheus-Moerman still in progress. A few issues have been cleared up and a few are still pending here: https://github.com/danielhatton/eights/issues
