### 2018-11-26 #974 @Kevin-Mattheus-Moerman
Thanks for the update @danielhatton, I would say that you should aim to have the improvements finished within a month to avoid loosing track of our reviewer.
