### 2018-12-23 danielhatton/eights#4 @danielhatton
@devinberg:

> Only difference I'm seeing when I run the examples is the positioning of the text in the title block.

Thanks, @devinberg .  This problem with the placement of fields in the title block in FreeCAD version 0.15 was described in the "bugs" section of the original Documentation.txt file as follows:

> If run under FreeCAD 0.15 (or earlier), the create_it method in the
> create_eights_drawing_sheet class incorrectly places the various items
> of text information correctly in the title block of the
> Drawing::FeaturePage object it creates.  That's because the number and
> order of elements, in the EditableTexts property of a
> Drawing::FeaturePage object using an ISO7200 template, changed between
> FreeCAD 0.15 and FreeCAD 0.16, and this version of the present module
> is tuned to place the text information correctly under FreeCAD 0.16.

However, I believe that, in the latest set of commits, I've fixed this bug (by having the eights.py module detect the FreeCAD version number under which it's running, and assign different title block fields to different array elements of the EditableTexts property depending on whether the version number is less than or greater than 0.155).  Therefore, also in the latest set of commits, I've deleted that paragraph from the "bugs" section of the Documentation.txt file.
