+++ README.md
@@ -1,0 +1,551 @@
+BibTeX: paper/paper  
+
+# About this documentation file
+
+This is file `README.md`
+
+`README.md` is the documentation for a software library centred
+on file `eights.py` (see "Description" section below).
+
+`README.md` was written by Dr. Daniel C. Hatton
+
+Copyright (C) 2017-2018 University of Plymouth Higher Education
+Corporation
+
+This program, including this documentation file, is free software: you
+can redistribute it and/or modify it under the terms of the GNU Lesser
+General Public License as published by the Free Software Foundation:
+version 3 of the License.
+
+This program is distributed in the hope that it will be useful, but
+WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+Lesser General Public License, and the GNU General Public License
+which it incorporates, for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License [in file `LICENSE`] along with this program.  If not, see
+<https://www.gnu.org/licenses/>.
+
+Daniel Hatton thanks Dr. Justin E. Rigden, Specialist Intellectual
+Property Advisor, for authorizing, on behalf of the University of
+Plymouth Higher Education Corporation, the release of this program
+under the licence terms stated above.
+
+# Contact
+
+Daniel Hatton can be contacted on <daniel.hatton@plymouth.ac.uk>
+or at
+Autonomous Marine Systems Group,
+School of Engineering,
+University of Plymouth,
+Plymouth,
+UK.
+PL4 8AA
+
+
+# Description
+
+File `eights.py` is a Python module intended for use with the
+open-source 3D computer-aided design package FreeCAD [@a:1:FVU].  The
+module automates the construction of a page of 2D axonometric drawings
+in first angle projection, in a style consistent (to the best of the
+module author's ability) with the BS 8888:2011 standard [@a:2011:TPD].
+
+# Statement of need
+
+The creative process behind detailed engineering design now typically
+takes place in 3D CAD packages [@Quintana:2010:WMB]. However, for this
+creative process actually to lead to an embodiment of the design being
+manufactured, this creative process must be followed by communication
+of the content of the design to a manufacturing facility. There are
+two necessary requirements for success: firstly, the communication
+must include information on all those aspects of the design that are
+necessary to complete manufacture, and to verify by inspection that
+the completed, manufactured artefact matches the design
+[@Quintana:2010:WMB]; secondly, the communication must be in a
+well-defined (usually graphical) language, so that it can be
+understood at the manufacturing facility in a way that permits
+successful manufacturing and inspection [@Quintana:2010:WMB;
+@Dobelis:2018:ICD]. The drive to meet both of these requirements has
+led to the development of published standards for design communication
+[@Dobelis:2018:ICD], of which BS 8888:2011 [@a:2011:TPD] is one. These
+standards are in a very advanced state of maturity for 2D engineering
+drawings, but despite efforts in recent years to develop similar
+standards for 3D models, the standards for 3D models remain somewhat
+less mature [@Quintana:2010:WMB]. As a result, 2D drawings remain a
+crucial medium for communication of design information to
+manufacturing facilities, and possessing the capability for automated
+generation of standards-compliant 2D drawings from a 3D model is an
+important criterion by which the technical quality of 3D CAD software
+packages is assessed [@Kannan:2008:MCD; @Hughes:2013:CW]. Hence, by
+automating the process of producing certain standards-compliant 2D
+drawings, the eights module offers the opportunity for the FreeCAD CAD
+package [@a:1:FVU] to be more favourably assessed against this
+criterion in future.
+
[…]
