+++ README.md
@@ -1,7 +1,3 @@
----
-bibliography: paper/paper.bib
----
-
 # About this documentation file
 
 This is file `README.md`
@@ -49,52 +45,55 @@ PL4 8AA
 # Description
 
 File `eights.py` is a Python module intended for use with the
-open-source 3D computer-aided design package FreeCAD [@a:1:FVU].  The
-module automates the construction of a page of 2D axonometric drawings
-in first angle projection, in a style consistent (to the best of the
-module author's ability) with the BS 8888:2011 standard [@a:2011:TPD].
+open-source 3D computer-aided design package FreeCAD (FreeCAD version
+0.15 user manual, n.d.).  The module automates the construction of a
+page of 2D axonometric drawings in first angle projection, in a style
+consistent (to the best of the module author's ability) with the BS
+8888:2011 standard (Technical product documentation and specification
+BS 8888, 2011).
 
 # Statement of need
 
 The creative process behind detailed engineering design now typically
-takes place in 3D CAD packages [@Quintana:2010:WMB]. However, for this
-creative process actually to lead to an embodiment of the design being
-manufactured, this creative process must be followed by communication
-of the content of the design to a manufacturing facility. There are
-two necessary requirements for success: firstly, the communication
-must include information on all those aspects of the design that are
-necessary to complete manufacture, and to verify by inspection that
-the completed, manufactured artefact matches the design
-[@Quintana:2010:WMB]; secondly, the communication must be in a
-well-defined (usually graphical) language, so that it can be
+takes place in 3D CAD packages (Quintana et al., 2010). However, for
+this creative process actually to lead to an embodiment of the design
+being manufactured, this creative process must be followed by
+communication of the content of the design to a manufacturing
+facility. There are two necessary requirements for success: firstly,
+the communication must include information on all those aspects of the
+design that are necessary to complete manufacture, and to verify by
+inspection that the completed, manufactured artefact matches the
+design (Quintana et al., 2010); secondly, the communication must be in
+a well-defined (usually graphical) language, so that it can be
 understood at the manufacturing facility in a way that permits
-successful manufacturing and inspection [@Quintana:2010:WMB;
-@Dobelis:2018:ICD]. The drive to meet both of these requirements has
-led to the development of published standards for design communication
-[@Dobelis:2018:ICD], of which BS 8888:2011 [@a:2011:TPD] is one. These
+successful manufacturing and inspection (Quintana et al., 2010;
+Dobelis et al., 2018). The drive to meet both of these requirements
+has led to the development of published standards for design
+communication (Dobelis et al., 2018), of which BS 8888:2011 (Technical
+product documentation and specification BS 8888, 2011) is one. These
 standards are in a very advanced state of maturity for 2D engineering
 drawings, but despite efforts in recent years to develop similar
 standards for 3D models, the standards for 3D models remain somewhat
-less mature [@Quintana:2010:WMB]. As a result, 2D drawings remain a
+less mature (Quintana et al., 2010). As a result, 2D drawings remain a
 crucial medium for communication of design information to
 manufacturing facilities, and possessing the capability for automated
 generation of standards-compliant 2D drawings from a 3D model is an
 important criterion by which the technical quality of 3D CAD software
-packages is assessed [@Kannan:2008:MCD; @Hughes:2013:CW]. Hence, by
+packages is assessed (Kannan & Vinay, 2008; Hughes, 2013). Hence, by
[…]
