### 2018-12-24 8c03cf6
Delete Documentation.txt
