### 2018-12-24 danielhatton/eights#4 @danielhatton
@devinberg :

> However, I think you could take your example documents and provide those with the intended output
> as a way of helping users test their system...
> This might be easier to do if you migrated your `documentation.txt` file to a `README.md` file. Then
> you could include some screenshots showing the intended output for a test input, providing users a
> comparison point.

In the latest set of commits, I have indeed replaced `documentation.txt` with a `README.md` file which includes embedded images of the intended outputs from the example scripts.
