### 2018-12-24 danielhatton/eights#4 @danielhatton
@devinberg

> I was able to use `eights.py` by placing it in the macros directory. However, I was not able to get it to
> work by saving the file in the working directory.

In the new `README.md` file, uploaded as part of the latest set of commits, I've clarified the "Installation" section to take account of this information.
