### 2018-12-24 danielhatton/eights#3 @danielhatton
In the latest commit, I've added a `CONTRIBUTING` file containing some appropriate guidance.  I've also created a new release, since I believe this completes addressing everything raised both here and in the "Installation and testing" issue thread.
