### 2018-12-24 danielhatton/eights#4 @danielhatton
@devinberg
> I've tried this on Windows 10 with both FreeCAD 0.17 and 0.15. Version 0.17 doesn't appear to work
> completely as intended, but 0.15 I think does.

Given what we both said later in the thread about the placement of fields in the title block, I suspect it might have been the other way round, and that in any case, the difference will have gone away in the new release.  I've just run all three example scripts on FreeCAD 0.17 under Fedora 29, and the results were as intended.
