### 2018-12-24 danielhatton/eights#4 @danielhatton
@danielhatton
> I've made copies (via "print to file" in the FreeCAD GUI) of the outputs I get from Single_part.py and
> Assembly.py, but every time I try to attach them to this comment, I get a message from GitHub
> saying 'Something went really wrong, and we can't process that file'. I'll try it again tomorrow, from a
> different network environment.

Just in case anyone else is trying to debug a similar error message to this one, I've been investigating its origin, and it appears to result from the fact that the "family filter" on my home ADSL connection, as currently configured, blocks access to anything in the domain `s3.amazonaws.com`; it looks like, without access to that domain, it's not possible either to upload or to download files attached to comments in issue threads on GitHub.  I can work around it by simply doing my uploads and downloads from a different network environment.
