### 2018-12-26 #974 @Kevin-Mattheus-Moerman
@danielhatton can I help? You can use `@whedon generate pdf` if you like
