### 2018-12-26 danielhatton/eights#4 @devinberg
@danielhatton

> Given what we both said later in the thread about the placement of fields in the title block, I suspect it might have been the other way round, and that in any case, the difference will have gone away in the new release.

I actually had trouble getting it to run at all on Windows 10 using FreeCAD 0.17 but I couldn't say with certainty that it wasn't an issue with my setup. Also, since your Prerequisites specify you've only tested version 0.15 on Windows, I don't think this is a big deal as far as the JOSS review is concerned. Thanks for your edits in the latest release.
