### 2018-12-26 #974 @devinberg
@Kevin-Mattheus-Moerman @danielhatton With the most recent edits and latest release, I feel that the Review Checklist is now satisfied.
