### 2019-01-03 #974 @Kevin-Mattheus-Moerman
@danielhatton congrats I will now recommend that this work is accepted in JOSS for publication. To proceed please do the following:

1. Archive a release of reviewed software using a service like Zenodo (or FigShare or similar, see also https://guides.github.com/activities/citable-code/)
Once you have a DOI of the archived version please list it here.
2. Please inform us of the version tag for the release that was archived. Should it be v0.1 by now?

Thanks.
