### 2019-01-03 danielhatton/eights#4 @danielhatton
@devinberg

> I actually had trouble getting it to run at all on Windows 10 using FreeCAD 0.17

Earlier today, I installed FreeCAD 0.17 on a Windows 10 machine, and I was able to get all three example scripts for `eights` to run correctly, but only after I'd restarted FreeCAD a couple of times to get it to recognise that there was a module called `eights` in its macros directory.  `README.md` already mentions that a restart of FreeCAD is sometimes needed to complete the installation of the `eights` module, but I think I'll add a comment to the effect that it sometimes needs to be restarted twice.
