### 2019-01-04 #974 @danielhatton
@Kevin-Mattheus-Moerman

> congrats I will now recommend that this work is accepted in JOSS for publication.

Brilliant, thank you very much.

> To proceed please do the following:
>     1. Archive a release of reviewed software using a service like Zenodo (or FigShare or similar, see > also https://guides.github.com/activities/citable-code/)
>        Once you have a DOI of the archived version please list it here.
>     2. Please inform us of the version tag for the release that was archived.

Will do.

> Should it be v0.1 by now?

The release of which @devinberg said 'the Review Checklist is now satisfied' was 0.1; but there is now a release 0.2, in which there's been a further clarification to one sentence in the "installation" section of `README.md`.  I think I'll archive and get the DOI for release 0.2.
