### 2019-01-04 #974 @danielhatton
Excellent.  Release 0.2 of `eights` now has DOI `10.5281/zenodo.2531603`.
