### 2019-01-20 #974 @labarba
Congratulations, @danielhatton, your paper is published!

Heartfelt thanks to the handling editor, @Kevin-Mattheus-Moerman, and the reviewer, @devinberg 🙏
